requirements.txt
Flask
torch
torchvision
opencv-python
Pillow
ultralytics
